Change jsx to tsx in EventDefinition files
