+++ graylog2-web-interface/src/components/event-definitions/event-definition-form/EventDetailsForm.tsx
@@ -36,7 +36,7 @@ type Props = {
   eventDefinition: EventDefinition,
   validation: {
     errors: {
-      title: string,
+      title?: string,
     }
   },
   onChange: (name: string, value: string | number) => void,
