+++ graylog2-web-interface/src/components/event-definitions/event-definition-form/FieldsForm.tsx
@@ -40,7 +40,7 @@ type Props = {
   eventDefinition: EventDefinition,
   validation: {
     errors: {
-      title: string,
+      title?: string,
     }
   },
   onChange: (name: string, value: any) => void,
